fix flag check
