+++ ethereum/swap_factory.go
@@ -44,7 +44,7 @@ type SwapFactorySwap struct {
 // SwapFactoryMetaData contains all meta data concerning the SwapFactory contract.
 var SwapFactoryMetaData = &bind.MetaData{
 	ABI: "[{\"inputs\":[{\"internalType\":\"address\",\"name\":\"trustedForwarder\",\"type\":\"address\"}],\"stateMutability\":\"nonpayable\",\"type\":\"constructor\"},{\"anonymous\":false,\"inputs\":[{\"indexed\":false,\"internalType\":\"bytes32\",\"name\":\"swapID\",\"type\":\"bytes32\"},{\"indexed\":false,\"internalType\":\"bytes32\",\"name\":\"s\",\"type\":\"bytes32\"}],\"name\":\"Claimed\",\"type\":\"event\"},{\"anonymous\":false,\"inputs\":[{\"indexed\":false,\"internalType\":\"bytes32\",\"name\":\"swapID\",\"type\":\"bytes32\"},{\"indexed\":false,\"internalType\":\"bytes32\",\"name\":\"claimKey\",\"type\":\"bytes32\"},{\"indexed\":false,\"internalType\":\"bytes32\",\"name\":\"refundKey\",\"type\":\"bytes32\"},{\"indexed\":false,\"internalType\":\"uint256\",\"name\":\"timeout_0\",\"type\":\"uint256\"},{\"indexed\":false,\"internalType\":\"uint256\",\"name\":\"timeout_1\",\"type\":\"uint256\"},{\"indexed\":false,\"internalType\":\"address\",\"name\":\"asset\",\"type\":\"address\"},{\"indexed\":false,\"internalType\":\"uint256\",\"name\":\"value\",\"type\":\"uint256\"}],\"name\":\"New\",\"type\":\"event\"},{\"anonymous\":false,\"inputs\":[{\"indexed\":false,\"internalType\":\"bytes32\",\"name\":\"swapID\",\"type\":\"bytes32\"}],\"name\":\"Ready\",\"type\":\"event\"},{\"anonymous\":false,\"inputs\":[{\"indexed\":false,\"internalType\":\"bytes32\",\"name\":\"swapID\",\"type\":\"bytes32\"},{\"indexed\":false,\"internalType\":\"bytes32\",\"name\":\"s\",\"type\":\"bytes32\"}],\"name\":\"Refunded\",\"type\":\"event\"},{\"inputs\":[],\"name\":\"_trustedForwarder\",\"outputs\":[{\"internalType\":\"address\",\"name\":\"\",\"type\":\"address\"}],\"stateMutability\":\"view\",\"type\":\"function\"},{\"inputs\":[{\"components\":[{\"internalType\":\"addresspayable\",\"name\":\"owner\",\"type\":\"address\"},{\"internalType\":\"addresspayable\",\"name\":\"claimer\",\"type\":\"address\"},{\"internalType\":\"bytes32\",\"name\":\"pubKeyClaim\",\"type\":\"bytes32\"},{\"internalType\":\"bytes32\",\"name\":\"pubKeyRefund\",\"type\":\"bytes32\"},{\"internalType\":\"uint256\",\"name\":\"timeout_0\",\"type\":\"uint256\"},{\"internalType\":\"uint256\",\"name\":\"timeout_1\",\"type\":\"uint256\"},{\"internalType\":\"address\",\"name\":\"asset\",\"type\":\"address\"},{\"internalType\":\"uint256\",\"name\":\"value\",\"type\":\"uint256\"},{\"internalType\":\"uint256\",\"name\":\"nonce\",\"type\":\"uint256\"}],\"internalType\":\"structSwapFactory.Swap\",\"name\":\"_swap\",\"type\":\"tuple\"},{\"internalType\":\"bytes32\",\"name\":\"_s\",\"type\":\"bytes32\"}],\"name\":\"claim\",\"outputs\":[],\"stateMutability\":\"nonpayable\",\"type\":\"function\"},{\"inputs\":[{\"components\":[{\"internalType\":\"addresspayable\",\"name\":\"owner\",\"type\":\"address\"},{\"internalType\":\"addresspayable\",\"name\":\"claimer\",\"type\":\"address\"},{\"internalType\":\"bytes32\",\"name\":\"pubKeyClaim\",\"type\":\"bytes32\"},{\"internalType\":\"bytes32\",\"name\":\"pubKeyRefund\",\"type\":\"bytes32\"},{\"internalType\":\"uint256\",\"name\":\"timeout_0\",\"type\":\"uint256\"},{\"internalType\":\"uint256\",\"name\":\"timeout_1\",\"type\":\"uint256\"},{\"internalType\":\"address\",\"name\":\"asset\",\"type\":\"address\"},{\"internalType\":\"uint256\",\"name\":\"value\",\"type\":\"uint256\"},{\"internalType\":\"uint256\",\"name\":\"nonce\",\"type\":\"uint256\"}],\"internalType\":\"structSwapFactory.Swap\",\"name\":\"_swap\",\"type\":\"tuple\"},{\"internalType\":\"bytes32\",\"name\":\"_s\",\"type\":\"bytes32\"},{\"internalType\":\"uint256\",\"name\":\"fee\",\"type\":\"uint256\"}],\"name\":\"claimRelayer\",\"outputs\":[],\"stateMutability\":\"nonpayable\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"address\",\"name\":\"forwarder\",\"type\":\"address\"}],\"name\":\"isTrustedForwarder\",\"outputs\":[{\"internalType\":\"bool\",\"name\":\"\",\"type\":\"bool\"}],\"stateMutability\":\"view\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"uint256\",\"name\":\"scalar\",\"type\":\"uint256\"},{\"internalType\":\"uint256\",\"name\":\"qKeccak\",\"type\":\"uint256\"}],\"name\":\"mulVerify\",\"outputs\":[{\"internalType\":\"bool\",\"name\":\"\",\"type\":\"bool\"}],\"stateMutability\":\"pure\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"bytes32\",\"name\":\"_pubKeyClaim\",\"type\":\"bytes32\"},{\"internalType\":\"bytes32\",\"name\":\"_pubKeyRefund\",\"type\":\"bytes32\"},{\"internalType\":\"addresspayable\",\"name\":\"_claimer\",\"type\":\"address\"},{\"internalType\":\"uint256\",\"name\":\"_timeoutDuration\",\"type\":\"uint256\"},{\"internalType\":\"address\",\"name\":\"_asset\",\"type\":\"address\"},{\"internalType\":\"uint256\",\"name\":\"_value\",\"type\":\"uint256\"},{\"internalType\":\"uint256\",\"name\":\"_nonce\",\"type\":\"uint256\"}],\"name\":\"newSwap\",\"outputs\":[{\"internalType\":\"bytes32\",\"name\":\"\",\"type\":\"bytes32\"}],\"stateMutability\":\"payable\",\"type\":\"function\"},{\"inputs\":[{\"components\":[{\"internalType\":\"addresspayable\",\"name\":\"owner\",\"type\":\"address\"},{\"internalType\":\"addresspayable\",\"name\":\"claimer\",\"type\":\"address\"},{\"internalType\":\"bytes32\",\"name\":\"pubKeyClaim\",\"type\":\"bytes32\"},{\"internalType\":\"bytes32\",\"name\":\"pubKeyRefund\",\"type\":\"bytes32\"},{\"internalType\":\"uint256\",\"name\":\"timeout_0\",\"type\":\"uint256\"},{\"internalType\":\"uint256\",\"name\":\"timeout_1\",\"type\":\"uint256\"},{\"internalType\":\"address\",\"name\":\"asset\",\"type\":\"address\"},{\"internalType\":\"uint256\",\"name\":\"value\",\"type\":\"uint256\"},{\"internalType\":\"uint256\",\"name\":\"nonce\",\"type\":\"uint256\"}],\"internalType\":\"structSwapFactory.Swap\",\"name\":\"_swap\",\"type\":\"tuple\"},{\"internalType\":\"bytes32\",\"name\":\"_s\",\"type\":\"bytes32\"}],\"name\":\"refund\",\"outputs\":[],\"stateMutability\":\"nonpayable\",\"type\":\"function\"},{\"inputs\":[{\"components\":[{\"internalType\":\"addresspayable\",\"name\":\"owner\",\"type\":\"address\"},{\"internalType\":\"addresspayable\",\"name\":\"claimer\",\"type\":\"address\"},{\"internalType\":\"bytes32\",\"name\":\"pubKeyClaim\",\"type\":\"bytes32\"},{\"internalType\":\"bytes32\",\"name\":\"pubKeyRefund\",\"type\":\"bytes32\"},{\"internalType\":\"uint256\",\"name\":\"timeout_0\",\"type\":\"uint256\"},{\"internalType\":\"uint256\",\"name\":\"timeout_1\",\"type\":\"uint256\"},{\"internalType\":\"address\",\"name\":\"asset\",\"type\":\"address\"},{\"internalType\":\"uint256\",\"name\":\"value\",\"type\":\"uint256\"},{\"internalType\":\"uint256\",\"name\":\"nonce\",\"type\":\"uint256\"}],\"internalType\":\"structSwapFactory.Swap\",\"name\":\"_swap\",\"type\":\"tuple\"}],\"name\":\"setReady\",\"outputs\":[],\"stateMutability\":\"nonpayable\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"bytes32\",\"name\":\"\",\"type\":\"bytes32\"}],\"name\":\"swaps\",\"outputs\":[{\"internalType\":\"enumSwapFactory.Stage\",\"name\":\"\",\"type\":\"uint8\"}],\"stateMutability\":\"view\",\"type\":\"function\"}]",
-	Bin: "0x60a06040523480156200001157600080fd5b50604051620024f9380380620024f98339818101604052810190620000379190620000de565b808073ffffffffffffffffffffffffffffffffffffffff1660808173ffffffffffffffffffffffffffffffffffffffff1681525050505062000110565b600080fd5b600073ffffffffffffffffffffffffffffffffffffffff82169050919050565b6000620000a68262000079565b9050919050565b620000b88162000099565b8114620000c457600080fd5b50565b600081519050620000d881620000ad565b92915050565b600060208284031215620000f757620000f662000074565b5b60006200010784828501620000c7565b91505092915050565b6080516123c662000133600039600081816105c101526105e701526123c66000f3fe6080604052600436106100865760003560e01c806373e4771c1161005957806373e4771c14610145578063aa0f87251461016e578063b32d1b4f1461019e578063eb84e7f2146101db578063fcaf229c1461021857610086565b80631e6c5acc1461008b57806356c022bb146100b4578063572b6c05146100df5780635cb969161461011c575b600080fd5b34801561009757600080fd5b506100b260048036038101906100ad91906115aa565b610241565b005b3480156100c057600080fd5b506100c96105bf565b6040516100d691906115fb565b60405180910390f35b3480156100eb57600080fd5b5061010660048036038101906101019190611616565b6105e3565b604051610113919061165e565b60405180910390f35b34801561012857600080fd5b50610143600480360381019061013e91906115aa565b61063b565b005b34801561015157600080fd5b5061016c60048036038101906101679190611679565b610761565b005b610188600480360381019061018391906116cf565b610974565b6040516101959190611780565b60405180910390f35b3480156101aa57600080fd5b506101c560048036038101906101c0919061179b565b610cab565b6040516101d2919061165e565b60405180910390f35b3480156101e757600080fd5b5061020260048036038101906101fd91906117db565b610db0565b60405161020f919061187f565b60405180910390f35b34801561022457600080fd5b5061023f600480360381019061023a919061189a565b610dd0565b005b60008260405160200161025491906119bb565b604051602081830303815290604052805190602001209050600080600083815260200190815260200160002060009054906101000a900460ff1690506003808111156102a3576102a2611808565b5b8160038111156102b6576102b5611808565b5b141580156102e95750600060038111156102d3576102d2611808565b5b8160038111156102e6576102e5611808565b5b14155b610328576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161031f90611a34565b60405180910390fd5b836000015173ffffffffffffffffffffffffffffffffffffffff163373ffffffffffffffffffffffffffffffffffffffff161461039a576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161039190611ac6565b60405180910390fd5b8360a00151421015806103e157508360800151421080156103e05750600260038111156103ca576103c9611808565b5b8160038111156103dd576103dc611808565b5b14155b5b610420576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161041790611b58565b60405180910390fd5b61042e838560600151610f69565b7e7c875846b687732a7579c19bb1dade66cd14e9f4f809565e2b2b5e76c72b4f828460405161045e929190611b78565b60405180910390a1600360008084815260200190815260200160002060006101000a81548160ff0219169083600381111561049c5761049b611808565b5b0217905550600073ffffffffffffffffffffffffffffffffffffffff168460c0015173ffffffffffffffffffffffffffffffffffffffff160361052d57836000015173ffffffffffffffffffffffffffffffffffffffff166108fc8560e001519081150290604051600060405180830381858888f19350505050158015610527573d6000803e3d6000fd5b506105b9565b8360c0015173ffffffffffffffffffffffffffffffffffffffff1663a9059cbb85600001518660e001516040518363ffffffff1660e01b8152600401610574929190611c0f565b6020604051808303816000875af1158015610593573d6000803e3d6000fd5b505050506040513d601f19601f820116820180604052508101906105b79190611c64565b505b50505050565b7f000000000000000000000000000000000000000000000000000000000000000081565b60007f000000000000000000000000000000000000000000000000000000000000000073ffffffffffffffffffffffffffffffffffffffff168273ffffffffffffffffffffffffffffffffffffffff16149050919050565b6106458282610fbc565b600073ffffffffffffffffffffffffffffffffffffffff168260c0015173ffffffffffffffffffffffffffffffffffffffff16036106d157816020015173ffffffffffffffffffffffffffffffffffffffff166108fc8360e001519081150290604051600060405180830381858888f193505050501580156106cb573d6000803e3d6000fd5b5061075d565b8160c0015173ffffffffffffffffffffffffffffffffffffffff1663a9059cbb83602001518460e001516040518363ffffffff1660e01b8152600401610718929190611c0f565b6020604051808303816000875af1158015610737573d6000803e3d6000fd5b505050506040513d601f19601f8201168201806040525081019061075b9190611c64565b505b5050565b61076a336105e3565b6107a9576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016107a090611d03565b60405180910390fd5b6107b38383610fbc565b600073ffffffffffffffffffffffffffffffffffffffff168360c0015173ffffffffffffffffffffffffffffffffffffffff160361089157826020015173ffffffffffffffffffffffffffffffffffffffff166108fc828560e001516108199190611d52565b9081150290604051600060405180830381858888f19350505050158015610844573d6000803e3d6000fd5b503273ffffffffffffffffffffffffffffffffffffffff166108fc829081150290604051600060405180830381858888f1935050505015801561088b573d6000803e3d6000fd5b5061096f565b8260c0015173ffffffffffffffffffffffffffffffffffffffff1663a9059cbb8460200151838660e001516108c69190611d52565b6040518363ffffffff1660e01b81526004016108e3929190611c0f565b6020604051808303816000875af1158015610902573d6000803e3d6000fd5b505050506040513d601f19601f820116820180604052508101906109269190611c64565b503273ffffffffffffffffffffffffffffffffffffffff166108fc829081150290604051600060405180830381858888f1935050505015801561096d573d6000803e3d6000fd5b505b505050565b600061097e61128e565b33816000019073ffffffffffffffffffffffffffffffffffffffff16908173ffffffffffffffffffffffffffffffffffffffff1681525050888160400181815250508781606001818152505086816020019073ffffffffffffffffffffffffffffffffffffffff16908173ffffffffffffffffffffffffffffffffffffffff16815250508542610a0e9190611d86565b816080018181525050600286610a249190611dba565b42610a2f9190611d86565b8160a0018181525050848160c0019073ffffffffffffffffffffffffffffffffffffffff16908173ffffffffffffffffffffffffffffffffffffffff1681525050838160e0018181525050600073ffffffffffffffffffffffffffffffffffffffff168160c0015173ffffffffffffffffffffffffffffffffffffffff1603610afd57348160e0015114610af8576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401610aef90611e86565b60405180910390fd5b610b87565b8060c0015173ffffffffffffffffffffffffffffffffffffffff166323b872dd33308460e001516040518463ffffffff1660e01b8152600401610b4293929190611ea6565b6020604051808303816000875af1158015610b61573d6000803e3d6000fd5b505050506040513d601f19601f82011682018060405250810190610b859190611c64565b505b8281610100018181525050600081604051602001610ba591906119bb565b60405160208183030381529060405280519060200120905060006003811115610bd157610bd0611808565b5b60008083815260200190815260200160002060009054906101000a900460ff166003811115610c0357610c02611808565b5b14610c0d57600080fd5b7f91446ce035ac29998b5473504609a5ef5e961005daba4630a1684b63be848f56818b8b85608001518660a001518760c001518860e00151604051610c589796959493929190611edd565b60405180910390a1600160008083815260200190815260200160002060006101000a81548160ff02191690836003811115610c9657610c95611808565b5b02179055508092505050979650505050505050565b60008060016000601b7f79be667ef9dcbbac55a06295ce870b07029bfcdb2dce28d959f2815b16f8179860001b7ffffffffffffffffffffffffffffffffebaaedce6af48a03bbfd25e8cd036414180610d0757610d06611f4c565b5b7f79be667ef9dcbbac55a06295ce870b07029bfcdb2dce28d959f2815b16f81798890960001b60405160008152602001604052604051610d4a949392919061200b565b6020604051602081039080840390855afa158015610d6c573d6000803e3d6000fd5b5050506020604051035190508073ffffffffffffffffffffffffffffffffffffffff168373ffffffffffffffffffffffffffffffffffffffff161491505092915050565b60006020528060005260406000206000915054906101000a900460ff1681565b600081604051602001610de391906119bb565b60405160208183030381529060405280519060200120905060016003811115610e0f57610e0e611808565b5b60008083815260200190815260200160002060009054906101000a900460ff166003811115610e4157610e40611808565b5b14610e81576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401610e789061209c565b60405180910390fd5b3373ffffffffffffffffffffffffffffffffffffffff16826000015173ffffffffffffffffffffffffffffffffffffffff1614610ef3576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401610eea9061212e565b60405180910390fd5b600260008083815260200190815260200160002060006101000a81548160ff02191690836003811115610f2957610f28611808565b5b02179055507f5fc23b25552757626e08b316cc2387ad1bc70ee1594af7204db4ce0c39f5d15f81604051610f5d9190611780565b60405180910390a15050565b610f798260001c8260001c610cab565b610fb8576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401610faf906121c0565b60405180910390fd5b5050565b600082604051602001610fcf91906119bb565b604051602081830303815290604052805190602001209050600080600083815260200190815260200160002060009054906101000a900460ff1690506000600381111561101f5761101e611808565b5b81600381111561103257611031611808565b5b03611072576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016110699061222c565b60405180910390fd5b60038081111561108557611084611808565b5b81600381111561109857611097611808565b5b036110d8576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016110cf90611a34565b60405180910390fd5b836020015173ffffffffffffffffffffffffffffffffffffffff163373ffffffffffffffffffffffffffffffffffffffff161461114a576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161114190612298565b60405180910390fd5b83608001514210158061118157506002600381111561116c5761116b611808565b5b81600381111561117f5761117e611808565b5b145b6111c0576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016111b790612304565b60405180910390fd5b8360a001514210611206576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016111fd90612370565b60405180910390fd5b611214838560400151610f69565b7f38d6042dbdae8e73a7f6afbabd3fbe0873f9f5ed3cd71294591c3908c2e65fee8284604051611245929190611b78565b60405180910390a1600360008084815260200190815260200160002060006101000a81548160ff0219169083600381111561128357611282611808565b5b021790555050505050565b604051806101200160405280600073ffffffffffffffffffffffffffffffffffffffff168152602001600073ffffffffffffffffffffffffffffffffffffffff16815260200160008019168152602001600080191681526020016000815260200160008152602001600073ffffffffffffffffffffffffffffffffffffffff16815260200160008152602001600081525090565b6000604051905090565b600080fd5b600080fd5b6000601f19601f8301169050919050565b7f4e487b7100000000000000000000000000000000000000000000000000000000600052604160045260246000fd5b61137f82611336565b810181811067ffffffffffffffff8211171561139e5761139d611347565b5b80604052505050565b60006113b1611322565b90506113bd8282611376565b919050565b600073ffffffffffffffffffffffffffffffffffffffff82169050919050565b60006113ed826113c2565b9050919050565b6113fd816113e2565b811461140857600080fd5b50565b60008135905061141a816113f4565b92915050565b6000819050919050565b61143381611420565b811461143e57600080fd5b50565b6000813590506114508161142a565b92915050565b6000819050919050565b61146981611456565b811461147457600080fd5b50565b60008135905061148681611460565b92915050565b6000611497826113c2565b9050919050565b6114a78161148c565b81146114b257600080fd5b50565b6000813590506114c48161149e565b92915050565b600061012082840312156114e1576114e0611331565b5b6114ec6101206113a7565b905060006114fc8482850161140b565b60008301525060206115108482850161140b565b602083015250604061152484828501611441565b604083015250606061153884828501611441565b606083015250608061154c84828501611477565b60808301525060a061156084828501611477565b60a08301525060c0611574848285016114b5565b60c08301525060e061158884828501611477565b60e08301525061010061159d84828501611477565b6101008301525092915050565b60008061014083850312156115c2576115c161132c565b5b60006115d0858286016114ca565b9250506101206115e285828601611441565b9150509250929050565b6115f58161148c565b82525050565b600060208201905061161060008301846115ec565b92915050565b60006020828403121561162c5761162b61132c565b5b600061163a848285016114b5565b91505092915050565b60008115159050919050565b61165881611643565b82525050565b6000602082019050611673600083018461164f565b92915050565b600080600061016084860312156116935761169261132c565b5b60006116a1868287016114ca565b9350506101206116b386828701611441565b9250506101406116c586828701611477565b9150509250925092565b600080600080600080600060e0888a0312156116ee576116ed61132c565b5b60006116fc8a828b01611441565b975050602061170d8a828b01611441565b965050604061171e8a828b0161140b565b955050606061172f8a828b01611477565b94505060806117408a828b016114b5565b93505060a06117518a828b01611477565b92505060c06117628a828b01611477565b91505092959891949750929550565b61177a81611420565b82525050565b60006020820190506117956000830184611771565b92915050565b600080604083850312156117b2576117b161132c565b5b60006117c085828601611477565b92505060206117d185828601611477565b9150509250929050565b6000602082840312156117f1576117f061132c565b5b60006117ff84828501611441565b91505092915050565b7f4e487b7100000000000000000000000000000000000000000000000000000000600052602160045260246000fd5b6004811061184857611847611808565b5b50565b600081905061185982611837565b919050565b60006118698261184b565b9050919050565b6118798161185e565b82525050565b60006020820190506118946000830184611870565b92915050565b600061012082840312156118b1576118b061132c565b5b60006118bf848285016114ca565b91505092915050565b6118d1816113e2565b82525050565b6118e081611420565b82525050565b6118ef81611456565b82525050565b6118fe8161148c565b82525050565b6101208201600082015161191b60008501826118c8565b50602082015161192e60208501826118c8565b50604082015161194160408501826118d7565b50606082015161195460608501826118d7565b50608082015161196760808501826118e6565b5060a082015161197a60a08501826118e6565b5060c082015161198d60c08501826118f5565b5060e08201516119a060e08501826118e6565b506101008201516119b56101008501826118e6565b50505050565b6000610120820190506119d16000830184611904565b92915050565b600082825260208201905092915050565b7f7377617020697320616c726561647920636f6d706c6574656400000000000000600082015250565b6000611a1e6019836119d7565b9150611a29826119e8565b602082019050919050565b60006020820190508181036000830152611a4d81611a11565b9050919050565b7f726566756e64206d7573742062652063616c6c6564206279207468652073776160008201527f70206f776e657200000000000000000000000000000000000000000000000000602082015250565b6000611ab06027836119d7565b9150611abb82611a54565b604082019050919050565b60006020820190508181036000830152611adf81611aa3565b9050919050565b7f697427732074686520636f756e74657270617274792773207475726e2c20756e60008201527f61626c6520746f20726566756e642c2074727920616761696e206c6174657200602082015250565b6000611b42603f836119d7565b9150611b4d82611ae6565b604082019050919050565b60006020820190508181036000830152611b7181611b35565b9050919050565b6000604082019050611b8d6000830185611771565b611b9a6020830184611771565b9392505050565b6000819050919050565b6000611bc6611bc1611bbc846113c2565b611ba1565b6113c2565b9050919050565b6000611bd882611bab565b9050919050565b6000611bea82611bcd565b9050919050565b611bfa81611bdf565b82525050565b611c0981611456565b82525050565b6000604082019050611c246000830185611bf1565b611c316020830184611c00565b9392505050565b611c4181611643565b8114611c4c57600080fd5b50565b600081519050611c5e81611c38565b92915050565b600060208284031215611c7a57611c7961132c565b5b6000611c8884828501611c4f565b91505092915050565b7f636c61696d52656c617965722063616e206f6e6c792062652063616c6c65642060008201527f62792061207472757374656420666f7277617264657200000000000000000000602082015250565b6000611ced6036836119d7565b9150611cf882611c91565b604082019050919050565b60006020820190508181036000830152611d1c81611ce0565b9050919050565b7f4e487b7100000000000000000000000000000000000000000000000000000000600052601160045260246000fd5b6000611d5d82611456565b9150611d6883611456565b9250828203905081811115611d8057611d7f611d23565b5b92915050565b6000611d9182611456565b9150611d9c83611456565b9250828201905080821115611db457611db3611d23565b5b92915050565b6000611dc582611456565b9150611dd083611456565b9250817fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff0483118215151615611e0957611e08611d23565b5b828202905092915050565b7f76616c7565206e6f742073616d652061732045544820616d6f756e742073656e60008201527f7400000000000000000000000000000000000000000000000000000000000000602082015250565b6000611e706021836119d7565b9150611e7b82611e14565b604082019050919050565b60006020820190508181036000830152611e9f81611e63565b9050919050565b6000606082019050611ebb60008301866115ec565b611ec860208301856115ec565b611ed56040830184611c00565b949350505050565b600060e082019050611ef2600083018a611771565b611eff6020830189611771565b611f0c6040830188611771565b611f196060830187611c00565b611f266080830186611c00565b611f3360a08301856115ec565b611f4060c0830184611c00565b98975050505050505050565b7f4e487b7100000000000000000000000000000000000000000000000000000000600052601260045260246000fd5b6000819050919050565b60008160001b9050919050565b6000611fad611fa8611fa384611f7b565b611f85565b611420565b9050919050565b611fbd81611f92565b82525050565b6000819050919050565b600060ff82169050919050565b6000611ff5611ff0611feb84611fc3565b611ba1565b611fcd565b9050919050565b61200581611fda565b82525050565b60006080820190506120206000830187611fb4565b61202d6020830186611ffc565b61203a6040830185611771565b6120476060830184611771565b95945050505050565b7f73776170206973206e6f7420696e2050454e44494e4720737461746500000000600082015250565b6000612086601c836119d7565b915061209182612050565b602082019050919050565b600060208201905081810360008301526120b581612079565b9050919050565b7f6f6e6c79207468652073776170206f776e65722063616e2063616c6c2073657460008201527f5265616479000000000000000000000000000000000000000000000000000000602082015250565b60006121186025836119d7565b9150612123826120bc565b604082019050919050565b600060208201905081810360008301526121478161210b565b9050919050565b7f70726f76696465642073656372657420646f6573206e6f74206d61746368207460008201527f6865206578706563746564207075626c6963206b657900000000000000000000602082015250565b60006121aa6036836119d7565b91506121b58261214e565b604082019050919050565b600060208201905081810360008301526121d98161219d565b9050919050565b7f696e76616c696420737761700000000000000000000000000000000000000000600082015250565b6000612216600c836119d7565b9150612221826121e0565b602082019050919050565b6000602082019050818103600083015261224581612209565b9050919050565b7f6f6e6c7920636c61696d65722063616e20636c61696d21000000000000000000600082015250565b60006122826017836119d7565b915061228d8261224c565b602082019050919050565b600060208201905081810360008301526122b181612275565b9050919050565b7f746f6f206561726c7920746f20636c61696d2100000000000000000000000000600082015250565b60006122ee6013836119d7565b91506122f9826122b8565b602082019050919050565b6000602082019050818103600083015261231d816122e1565b9050919050565b7f746f6f206c61746520746f20636c61696d210000000000000000000000000000600082015250565b600061235a6012836119d7565b915061236582612324565b602082019050919050565b600060208201905081810360008301526123898161234d565b905091905056fea2646970667358221220bab6e7255a86895a5f1fa7bdd9c2819bf4eb6bb22c6a72104531fb4397c1354e64736f6c63430008100033",
+	Bin: "0x60a06040523480156200001157600080fd5b50604051620024e1380380620024e18339818101604052810190620000379190620000de565b808073ffffffffffffffffffffffffffffffffffffffff1660808173ffffffffffffffffffffffffffffffffffffffff1681525050505062000110565b600080fd5b600073ffffffffffffffffffffffffffffffffffffffff82169050919050565b6000620000a68262000079565b9050919050565b620000b88162000099565b8114620000c457600080fd5b50565b600081519050620000d881620000ad565b92915050565b600060208284031215620000f757620000f662000074565b5b60006200010784828501620000c7565b91505092915050565b6080516123ae62000133600039600081816105c101526105e701526123ae6000f3fe6080604052600436106100865760003560e01c806373e4771c1161005957806373e4771c14610145578063aa0f87251461016e578063b32d1b4f1461019e578063eb84e7f2146101db578063fcaf229c1461021857610086565b80631e6c5acc1461008b57806356c022bb146100b4578063572b6c05146100df5780635cb969161461011c575b600080fd5b34801561009757600080fd5b506100b260048036038101906100ad91906115aa565b610241565b005b3480156100c057600080fd5b506100c96105bf565b6040516100d691906115fb565b60405180910390f35b3480156100eb57600080fd5b5061010660048036038101906101019190611616565b6105e3565b604051610113919061165e565b60405180910390f35b34801561012857600080fd5b50610143600480360381019061013e91906115aa565b61063b565b005b34801561015157600080fd5b5061016c60048036038101906101679190611679565b610761565b005b610188600480360381019061018391906116cf565b610974565b6040516101959190611780565b60405180910390f35b3480156101aa57600080fd5b506101c560048036038101906101c0919061179b565b610cab565b6040516101d2919061165e565b60405180910390f35b3480156101e757600080fd5b5061020260048036038101906101fd91906117db565b610db0565b60405161020f919061187f565b60405180910390f35b34801561022457600080fd5b5061023f600480360381019061023a919061189a565b610dd0565b005b60008260405160200161025491906119bb565b604051602081830303815290604052805190602001209050600080600083815260200190815260200160002060009054906101000a900460ff1690506003808111156102a3576102a2611808565b5b8160038111156102b6576102b5611808565b5b141580156102e95750600060038111156102d3576102d2611808565b5b8160038111156102e6576102e5611808565b5b14155b610328576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161031f90611a34565b60405180910390fd5b836000015173ffffffffffffffffffffffffffffffffffffffff163373ffffffffffffffffffffffffffffffffffffffff161461039a576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161039190611ac6565b60405180910390fd5b8360a00151421015806103e157508360800151421080156103e05750600260038111156103ca576103c9611808565b5b8160038111156103dd576103dc611808565b5b14155b5b610420576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161041790611b58565b60405180910390fd5b61042e838560600151610f69565b7e7c875846b687732a7579c19bb1dade66cd14e9f4f809565e2b2b5e76c72b4f828460405161045e929190611b78565b60405180910390a1600360008084815260200190815260200160002060006101000a81548160ff0219169083600381111561049c5761049b611808565b5b0217905550600073ffffffffffffffffffffffffffffffffffffffff168460c0015173ffffffffffffffffffffffffffffffffffffffff160361052d57836000015173ffffffffffffffffffffffffffffffffffffffff166108fc8560e001519081150290604051600060405180830381858888f19350505050158015610527573d6000803e3d6000fd5b506105b9565b8360c0015173ffffffffffffffffffffffffffffffffffffffff1663a9059cbb85600001518660e001516040518363ffffffff1660e01b8152600401610574929190611c0f565b6020604051808303816000875af1158015610593573d6000803e3d6000fd5b505050506040513d601f19601f820116820180604052508101906105b79190611c64565b505b50505050565b7f000000000000000000000000000000000000000000000000000000000000000081565b60007f000000000000000000000000000000000000000000000000000000000000000073ffffffffffffffffffffffffffffffffffffffff168273ffffffffffffffffffffffffffffffffffffffff16149050919050565b6106458282610fbc565b600073ffffffffffffffffffffffffffffffffffffffff168260c0015173ffffffffffffffffffffffffffffffffffffffff16036106d157816020015173ffffffffffffffffffffffffffffffffffffffff166108fc8360e001519081150290604051600060405180830381858888f193505050501580156106cb573d6000803e3d6000fd5b5061075d565b8160c0015173ffffffffffffffffffffffffffffffffffffffff1663a9059cbb83602001518460e001516040518363ffffffff1660e01b8152600401610718929190611c0f565b6020604051808303816000875af1158015610737573d6000803e3d6000fd5b505050506040513d601f19601f8201168201806040525081019061075b9190611c64565b505b5050565b61076a336105e3565b6107a9576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016107a090611d03565b60405180910390fd5b6107b38383610fbc565b600073ffffffffffffffffffffffffffffffffffffffff168360c0015173ffffffffffffffffffffffffffffffffffffffff160361089157826020015173ffffffffffffffffffffffffffffffffffffffff166108fc828560e001516108199190611d52565b9081150290604051600060405180830381858888f19350505050158015610844573d6000803e3d6000fd5b503273ffffffffffffffffffffffffffffffffffffffff166108fc829081150290604051600060405180830381858888f1935050505015801561088b573d6000803e3d6000fd5b5061096f565b8260c0015173ffffffffffffffffffffffffffffffffffffffff1663a9059cbb8460200151838660e001516108c69190611d52565b6040518363ffffffff1660e01b81526004016108e3929190611c0f565b6020604051808303816000875af1158015610902573d6000803e3d6000fd5b505050506040513d601f19601f820116820180604052508101906109269190611c64565b503273ffffffffffffffffffffffffffffffffffffffff166108fc829081150290604051600060405180830381858888f1935050505015801561096d573d6000803e3d6000fd5b505b505050565b600061097e61128e565b33816000019073ffffffffffffffffffffffffffffffffffffffff16908173ffffffffffffffffffffffffffffffffffffffff1681525050888160400181815250508781606001818152505086816020019073ffffffffffffffffffffffffffffffffffffffff16908173ffffffffffffffffffffffffffffffffffffffff16815250508542610a0e9190611d86565b816080018181525050600286610a249190611dba565b42610a2f9190611d86565b8160a0018181525050848160c0019073ffffffffffffffffffffffffffffffffffffffff16908173ffffffffffffffffffffffffffffffffffffffff1681525050838160e0018181525050600073ffffffffffffffffffffffffffffffffffffffff168160c0015173ffffffffffffffffffffffffffffffffffffffff1603610afd57348160e0015114610af8576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401610aef90611e6e565b60405180910390fd5b610b87565b8060c0015173ffffffffffffffffffffffffffffffffffffffff166323b872dd33308460e001516040518463ffffffff1660e01b8152600401610b4293929190611e8e565b6020604051808303816000875af1158015610b61573d6000803e3d6000fd5b505050506040513d601f19601f82011682018060405250810190610b859190611c64565b505b8281610100018181525050600081604051602001610ba591906119bb565b60405160208183030381529060405280519060200120905060006003811115610bd157610bd0611808565b5b60008083815260200190815260200160002060009054906101000a900460ff166003811115610c0357610c02611808565b5b14610c0d57600080fd5b7f91446ce035ac29998b5473504609a5ef5e961005daba4630a1684b63be848f56818b8b85608001518660a001518760c001518860e00151604051610c589796959493929190611ec5565b60405180910390a1600160008083815260200190815260200160002060006101000a81548160ff02191690836003811115610c9657610c95611808565b5b02179055508092505050979650505050505050565b60008060016000601b7f79be667ef9dcbbac55a06295ce870b07029bfcdb2dce28d959f2815b16f8179860001b7ffffffffffffffffffffffffffffffffebaaedce6af48a03bbfd25e8cd036414180610d0757610d06611f34565b5b7f79be667ef9dcbbac55a06295ce870b07029bfcdb2dce28d959f2815b16f81798890960001b60405160008152602001604052604051610d4a9493929190611ff3565b6020604051602081039080840390855afa158015610d6c573d6000803e3d6000fd5b5050506020604051035190508073ffffffffffffffffffffffffffffffffffffffff168373ffffffffffffffffffffffffffffffffffffffff161491505092915050565b60006020528060005260406000206000915054906101000a900460ff1681565b600081604051602001610de391906119bb565b60405160208183030381529060405280519060200120905060016003811115610e0f57610e0e611808565b5b60008083815260200190815260200160002060009054906101000a900460ff166003811115610e4157610e40611808565b5b14610e81576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401610e7890612084565b60405180910390fd5b3373ffffffffffffffffffffffffffffffffffffffff16826000015173ffffffffffffffffffffffffffffffffffffffff1614610ef3576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401610eea90612116565b60405180910390fd5b600260008083815260200190815260200160002060006101000a81548160ff02191690836003811115610f2957610f28611808565b5b02179055507f5fc23b25552757626e08b316cc2387ad1bc70ee1594af7204db4ce0c39f5d15f81604051610f5d9190611780565b60405180910390a15050565b610f798260001c8260001c610cab565b610fb8576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401610faf906121a8565b60405180910390fd5b5050565b600082604051602001610fcf91906119bb565b604051602081830303815290604052805190602001209050600080600083815260200190815260200160002060009054906101000a900460ff1690506000600381111561101f5761101e611808565b5b81600381111561103257611031611808565b5b03611072576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161106990612214565b60405180910390fd5b60038081111561108557611084611808565b5b81600381111561109857611097611808565b5b036110d8576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016110cf90611a34565b60405180910390fd5b836020015173ffffffffffffffffffffffffffffffffffffffff163373ffffffffffffffffffffffffffffffffffffffff161461114a576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161114190612280565b60405180910390fd5b83608001514210158061118157506002600381111561116c5761116b611808565b5b81600381111561117f5761117e611808565b5b145b6111c0576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016111b7906122ec565b60405180910390fd5b8360a001514210611206576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016111fd90612358565b60405180910390fd5b611214838560400151610f69565b7f38d6042dbdae8e73a7f6afbabd3fbe0873f9f5ed3cd71294591c3908c2e65fee8284604051611245929190611b78565b60405180910390a1600360008084815260200190815260200160002060006101000a81548160ff0219169083600381111561128357611282611808565b5b021790555050505050565b604051806101200160405280600073ffffffffffffffffffffffffffffffffffffffff168152602001600073ffffffffffffffffffffffffffffffffffffffff16815260200160008019168152602001600080191681526020016000815260200160008152602001600073ffffffffffffffffffffffffffffffffffffffff16815260200160008152602001600081525090565b6000604051905090565b600080fd5b600080fd5b6000601f19601f8301169050919050565b7f4e487b7100000000000000000000000000000000000000000000000000000000600052604160045260246000fd5b61137f82611336565b810181811067ffffffffffffffff8211171561139e5761139d611347565b5b80604052505050565b60006113b1611322565b90506113bd8282611376565b919050565b600073ffffffffffffffffffffffffffffffffffffffff82169050919050565b60006113ed826113c2565b9050919050565b6113fd816113e2565b811461140857600080fd5b50565b60008135905061141a816113f4565b92915050565b6000819050919050565b61143381611420565b811461143e57600080fd5b50565b6000813590506114508161142a565b92915050565b6000819050919050565b61146981611456565b811461147457600080fd5b50565b60008135905061148681611460565b92915050565b6000611497826113c2565b9050919050565b6114a78161148c565b81146114b257600080fd5b50565b6000813590506114c48161149e565b92915050565b600061012082840312156114e1576114e0611331565b5b6114ec6101206113a7565b905060006114fc8482850161140b565b60008301525060206115108482850161140b565b602083015250604061152484828501611441565b604083015250606061153884828501611441565b606083015250608061154c84828501611477565b60808301525060a061156084828501611477565b60a08301525060c0611574848285016114b5565b60c08301525060e061158884828501611477565b60e08301525061010061159d84828501611477565b6101008301525092915050565b60008061014083850312156115c2576115c161132c565b5b60006115d0858286016114ca565b9250506101206115e285828601611441565b9150509250929050565b6115f58161148c565b82525050565b600060208201905061161060008301846115ec565b92915050565b60006020828403121561162c5761162b61132c565b5b600061163a848285016114b5565b91505092915050565b60008115159050919050565b61165881611643565b82525050565b6000602082019050611673600083018461164f565b92915050565b600080600061016084860312156116935761169261132c565b5b60006116a1868287016114ca565b9350506101206116b386828701611441565b9250506101406116c586828701611477565b9150509250925092565b600080600080600080600060e0888a0312156116ee576116ed61132c565b5b60006116fc8a828b01611441565b975050602061170d8a828b01611441565b965050604061171e8a828b0161140b565b955050606061172f8a828b01611477565b94505060806117408a828b016114b5565b93505060a06117518a828b01611477565b92505060c06117628a828b01611477565b91505092959891949750929550565b61177a81611420565b82525050565b60006020820190506117956000830184611771565b92915050565b600080604083850312156117b2576117b161132c565b5b60006117c085828601611477565b92505060206117d185828601611477565b9150509250929050565b6000602082840312156117f1576117f061132c565b5b60006117ff84828501611441565b91505092915050565b7f4e487b7100000000000000000000000000000000000000000000000000000000600052602160045260246000fd5b6004811061184857611847611808565b5b50565b600081905061185982611837565b919050565b60006118698261184b565b9050919050565b6118798161185e565b82525050565b60006020820190506118946000830184611870565b92915050565b600061012082840312156118b1576118b061132c565b5b60006118bf848285016114ca565b91505092915050565b6118d1816113e2565b82525050565b6118e081611420565b82525050565b6118ef81611456565b82525050565b6118fe8161148c565b82525050565b6101208201600082015161191b60008501826118c8565b50602082015161192e60208501826118c8565b50604082015161194160408501826118d7565b50606082015161195460608501826118d7565b50608082015161196760808501826118e6565b5060a082015161197a60a08501826118e6565b5060c082015161198d60c08501826118f5565b5060e08201516119a060e08501826118e6565b506101008201516119b56101008501826118e6565b50505050565b6000610120820190506119d16000830184611904565b92915050565b600082825260208201905092915050565b7f7377617020697320616c726561647920636f6d706c6574656400000000000000600082015250565b6000611a1e6019836119d7565b9150611a29826119e8565b602082019050919050565b60006020820190508181036000830152611a4d81611a11565b9050919050565b7f726566756e64206d7573742062652063616c6c6564206279207468652073776160008201527f70206f776e657200000000000000000000000000000000000000000000000000602082015250565b6000611ab06027836119d7565b9150611abb82611a54565b604082019050919050565b60006020820190508181036000830152611adf81611aa3565b9050919050565b7f697427732074686520636f756e74657270617274792773207475726e2c20756e60008201527f61626c6520746f20726566756e642c2074727920616761696e206c6174657200602082015250565b6000611b42603f836119d7565b9150611b4d82611ae6565b604082019050919050565b60006020820190508181036000830152611b7181611b35565b9050919050565b6000604082019050611b8d6000830185611771565b611b9a6020830184611771565b9392505050565b6000819050919050565b6000611bc6611bc1611bbc846113c2565b611ba1565b6113c2565b9050919050565b6000611bd882611bab565b9050919050565b6000611bea82611bcd565b9050919050565b611bfa81611bdf565b82525050565b611c0981611456565b82525050565b6000604082019050611c246000830185611bf1565b611c316020830184611c00565b9392505050565b611c4181611643565b8114611c4c57600080fd5b50565b600081519050611c5e81611c38565b92915050565b600060208284031215611c7a57611c7961132c565b5b6000611c8884828501611c4f565b91505092915050565b7f636c61696d52656c617965722063616e206f6e6c792062652063616c6c65642060008201527f62792061207472757374656420666f7277617264657200000000000000000000602082015250565b6000611ced6036836119d7565b9150611cf882611c91565b604082019050919050565b60006020820190508181036000830152611d1c81611ce0565b9050919050565b7f4e487b7100000000000000000000000000000000000000000000000000000000600052601160045260246000fd5b6000611d5d82611456565b9150611d6883611456565b9250828203905081811115611d8057611d7f611d23565b5b92915050565b6000611d9182611456565b9150611d9c83611456565b9250828201905080821115611db457611db3611d23565b5b92915050565b6000611dc582611456565b9150611dd083611456565b9250828202611dde81611456565b91508282048414831517611df557611df4611d23565b5b5092915050565b7f76616c7565206e6f742073616d652061732045544820616d6f756e742073656e60008201527f7400000000000000000000000000000000000000000000000000000000000000602082015250565b6000611e586021836119d7565b9150611e6382611dfc565b604082019050919050565b60006020820190508181036000830152611e8781611e4b565b9050919050565b6000606082019050611ea360008301866115ec565b611eb060208301856115ec565b611ebd6040830184611c00565b949350505050565b600060e082019050611eda600083018a611771565b611ee76020830189611771565b611ef46040830188611771565b611f016060830187611c00565b611f0e6080830186611c00565b611f1b60a08301856115ec565b611f2860c0830184611c00565b98975050505050505050565b7f4e487b7100000000000000000000000000000000000000000000000000000000600052601260045260246000fd5b6000819050919050565b60008160001b9050919050565b6000611f95611f90611f8b84611f63565b611f6d565b611420565b9050919050565b611fa581611f7a565b82525050565b6000819050919050565b600060ff82169050919050565b6000611fdd611fd8611fd384611fab565b611ba1565b611fb5565b9050919050565b611fed81611fc2565b82525050565b60006080820190506120086000830187611f9c565b6120156020830186611fe4565b6120226040830185611771565b61202f6060830184611771565b95945050505050565b7f73776170206973206e6f7420696e2050454e44494e4720737461746500000000600082015250565b600061206e601c836119d7565b915061207982612038565b602082019050919050565b6000602082019050818103600083015261209d81612061565b9050919050565b7f6f6e6c79207468652073776170206f776e65722063616e2063616c6c2073657460008201527f5265616479000000000000000000000000000000000000000000000000000000602082015250565b60006121006025836119d7565b915061210b826120a4565b604082019050919050565b6000602082019050818103600083015261212f816120f3565b9050919050565b7f70726f76696465642073656372657420646f6573206e6f74206d61746368207460008201527f6865206578706563746564207075626c6963206b657900000000000000000000602082015250565b60006121926036836119d7565b915061219d82612136565b604082019050919050565b600060208201905081810360008301526121c181612185565b9050919050565b7f696e76616c696420737761700000000000000000000000000000000000000000600082015250565b60006121fe600c836119d7565b9150612209826121c8565b602082019050919050565b6000602082019050818103600083015261222d816121f1565b9050919050565b7f6f6e6c7920636c61696d65722063616e20636c61696d21000000000000000000600082015250565b600061226a6017836119d7565b915061227582612234565b602082019050919050565b600060208201905081810360008301526122998161225d565b9050919050565b7f746f6f206561726c7920746f20636c61696d2100000000000000000000000000600082015250565b60006122d66013836119d7565b91506122e1826122a0565b602082019050919050565b60006020820190508181036000830152612305816122c9565b9050919050565b7f746f6f206c61746520746f20636c61696d210000000000000000000000000000600082015250565b60006123426012836119d7565b915061234d8261230c565b602082019050919050565b6000602082019050818103600083015261237181612335565b905091905056fea26469706673582212207cf3bfd05d4d227b0299c844617defd2a818d3d34ae3009caba922d7b9611f6c64736f6c63430008110033",
 }
 
 // SwapFactoryABI is the input ABI used to generate the binding from.
